Improve strings used in foreground service
From a user perspective this service isn't anything "in foreground", but
rather a background sync. Therefore don't mention "alert" or
"foreground" in the strings of this notification channel.

Also don't hardcode the notification title. "Event notification" sounds
like that this notification is some reminder of an event, but it isn't.

Change-Id: Iea2a67eb605afa6ce5c9f789252d47ed91df3638